Completed task 6

+++ 0x00-hello_world/6-size.c
@@ -9,7 +9,7 @@ int main(void)
 {
 printf("Size of a char: %d \n", (int)sizeof(char));
 printf("Size of an int: %d \n", (int)sizeof(int));
-printf("Size of a long int: %d \n", (int)sizeof(int));
+printf("Size of a long int: %d \n", (int)sizeof(long));
 printf("Size of a long long int: %d \n", (int)sizeof(long long int));
 printf("Size of a float: %d \n", (int)sizeof(float));
 return (0);
